+++ R/helper_functions.R
@@ -14,11 +14,13 @@ Assert <- function (statement) if (!statement) stop(deparse(statement), " is FAL
 #' @return edges in the format expected by \code{tree$edge},
 #'         where \code{tree} is a tree of class \code{phylo}.
 #' @keywords internal
+#' @export
 ListToMatrix <- function (edgeList) matrix(c(edgeList[[1]], edgeList[[2]]), ncol=2)
 
 #' Edge matrix to edge list
 #' @param edge edges in the matrix format used by \code{tree$edge}, where \code{tree} is a tree of class \code{phylo}
 #' @return tree edges in the format list(parent, child).
+#' @export
 MatrixToList <- function (edge) list(edge[, 1], edge[, 2])
 
 #' Descendant Edges
@@ -50,6 +52,22 @@ DescendantEdges <- function (edge, parent, child, nEdge = length(parent)) {
     ret[edge:nextEdge] <- TRUE 
     return(ret)
   }
+}
+
+#' All Descendant Edges
+#'
+#' @return a matrix of class logical, with row N specifying whether each edge is a descendant of edge N
+#'         (or the edge itself)
+#' @describeIn DescendantEdges Quickly identifies edges that are 'descended' from each edge in a tree
+AllDescendantEdges <- function (parent, child, nEdge = length(parent)) {
+  ret <- diag(nEdge) == 1
+  blankLogical <- logical(nEdge)
+  allEdges <- seq_len(nEdge)
+  for (edge in rev(allEdges[child > parent[1]])) {
+    ret[edge, ] <- apply(ret[parent == child[edge], ], 2, any)
+    ret[edge, edge] <- TRUE
+  }
+  ret
 }
 
 #' Ancestral edge