+++ man-roxygen/outgroupTipsParam.R
@@ -1,0 +1,2 @@
+#' @param outgroupTips Character vector specifying the names of the tips to 
+#'                     include in the outgroup.